Safari-609.1.11.1/LayoutTests/fast/workers/self-hasOwnProperty-expected.txt

Tests that self.hasOwnProperty() works in workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


self.hasOwnProperty('DedicatedWorkerGlobalScope'): true
self.hasOwnProperty('WorkerGlobalScope'): true
self.hasOwnProperty('postMessage'): true
self.hasOwnProperty('IDBDatabase'): true
PASS successfullyParsed is true

TEST COMPLETE
